+++ test/Sensorario/ValueObject/ValueObjectTest.php
@@ -14,15 +14,18 @@ namespace Sensorario\ValueObject;
 use DateInterval;
 use DateTime;
 use PHPUnit_Framework_TestCase;
+use Sensorario\Resources\Bar;
 use Sensorario\Resources\BirthDay;
+use Sensorario\Resources\Foo;
+use Sensorario\Resources\MandatoryDependency;
+use Sensorario\Resources\SomeApiRequest;
 use Sensorario\Services\ExportJSON;
-use Sensorario\ValueObject\Exception\InvalidKeyException;
-use Sensorario\ValueObject\Exception\UndefinedMandatoryPropertyException;
+use Sensorario\Services\PropertyType;
 
 final class ValueObjectTest extends PHPUnit_Framework_TestCase
 {
     /**
-     * @expectedException              Sensorario\ValueObject\Exceptions\InvalidMethodException
+     * @expectedException              RuntimeException
      * @expectedExceptionMessageRegExp #Method `.*::.*()` is not yet implemented#
      */
     public function testExceptionIsThrownWhenNotYetImplementedMethodIsCalled()
@@ -36,7 +39,7 @@ final class ValueObjectTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException              Sensorario\ValueObject\Exceptions\InvalidKeyException
+     * @expectedException              RuntimeException
      * @expectedExceptionMessageRegExp #Key `.*::.*` is not allowed#
      */
     public function testNotAllowedFieldThroghRuntimeException()
@@ -49,7 +52,7 @@ final class ValueObjectTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException              Sensorario\ValueObject\Exceptions\UndefinedMandatoryPropertyException
+     * @expectedException              RuntimeException
      * @expectedExceptionMessageRegExp #Property `.*::.*` is mandatory but not set#
      */
     public function testMissingMandatoryFieldThroghRuntimeException()
@@ -79,7 +82,7 @@ final class ValueObjectTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException Sensorario\ValueObject\Exceptions\InvalidFactoryMethodException
+     * @expectedException RuntimeException
      */
     public function testFactoryMethods()
     {
@@ -101,7 +104,7 @@ final class ValueObjectTest extends PHPUnit_Framework_TestCase
         $foo = Bar::box();
 
         $this->assertFalse(
-            $foo->propertyExists('nonExistentProperty')
+            $foo->hasProperty('nonExistentProperty')
         );
     }
 
@@ -128,7 +131,7 @@ final class ValueObjectTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException Sensorario\ValueObject\Exceptions\InvalidKeyOrValueException
+     * @expectedException RuntimeException
      */
     public function testThroughExceptionWhenNoValuesProvided()
     {
@@ -137,7 +140,7 @@ final class ValueObjectTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException        Sensorario\ValueObject\Exceptions\InvalidValueException
+     * @expectedException              RuntimeException
      * @expectedExceptionMessage Value `42` is not allowed for key `someApiParameter`
      */
     public function testAllowedValueForAField()
@@ -148,7 +151,7 @@ final class ValueObjectTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException              Sensorario\ValueObject\Exceptions\InvalidTypeException
+     * @expectedException              RuntimeException
      * @expectedExceptionMessageRegExp #Attribute `.*` must be an object#
      */
     public function testPropertyCouldBeAnObject()
@@ -159,7 +162,7 @@ final class ValueObjectTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException              Sensorario\ValueObject\Exceptions\InvalidTypeException
+     * @expectedException              RuntimeException
      * @expectedExceptionMessageRegExp #Attribute `.*` must be an object of type DateTime#
      */
     public function testPropertyCouldBeTheRightnObject()
@@ -167,30 +170,6 @@ final class ValueObjectTest extends PHPUnit_Framework_TestCase
         $birthday = BirthDay::box([
             'date' => new DateInterval('P1D'),
         ]);
-    }
-
-    public function testPropertiesTypeWhenObject()
-    {
-        $birthday = BirthDay::box([
-            'date' => new DateTime('2015'),
-        ]);
-
-        $this->assertEquals(
-            'DateTime',
-            $birthday->getPropertyType('date')
-        );
-    }
-
-    public function testPropertiesTypeWhenString()
-    {
-        $foo = Foo::box([
-            'name' => 'Simone'
-        ]);
-
-        $this->assertEquals(
-            'string',
-            $foo->getPropertyType('name')
-        );
     }
 
     public function testPropertiesAccessor()
@@ -207,7 +186,7 @@ final class ValueObjectTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException              Sensorario\ValueObject\Exceptions\UndefinedMandatoryPropertyException
+     * @expectedException              RuntimeException
      * @expectedExceptionMessageRegExp #Property `.*::.*` is mandatory but not set#
      */
     public function test()
@@ -216,101 +195,5 @@ final class ValueObjectTest extends PHPUnit_Framework_TestCase
             'foo' => 'bar',
             'world' => 'bar',
         ]);
-    }
-}
-
-final class MandatoryDependency extends ValueObject
-{
-    public static function mandatory()
-    {
-        return [
-            'foo',
-            'hello' => [
-                'if_present' => 'world',
-            ]
-        ];
-    }
-
-    public static function allowed()
-    {
-        return [
-            'hello',
-            'world',
-        ];
-    }
-}
-
-final class SomeApiRequest extends ValueObject
-{
-    public static function mandatory()
-    {
-        return [
-            'someApiParameter',
-        ];
-    }
-
-    public static function allowedValues()
-    {
-        return [
-            'someApiParameter' => [
-                'hello',
-                'world'
-            ],
-        ];
-    }
-}
-
-/**
- * Example class
- *
- * This kind of Vo provide one mandatory field, and two allowed
- */
-final class Foo extends ValueObject
-{
-    /**
-     * Only one mandatory field here
-     */
-    public static function mandatory()
-    {
-        return [
-            'name',
-        ];
-    }
-
-    /**
-     * This VO allows two fields
-     */
-    public static function allowed()
-    {
-        return [
-            'name',
-            'surname',
-        ];
-    }
-}
-
-/**
- * In this case, we have a default value
- */
-final class Bar extends ValueObject
-{
-    /**
-     * Only one mandatory field here
-     */
-    public static function allowed()
-    {
-        return [
-            'name',
-        ];
-    }
-
-    /**
-     * Only one default value
-     */
-    public static function defaults()
-    {
-        return [
-            'name' => 'Firefox',
-        ];
     }
 }