+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 # Release 1.0
 
+* [closed #6] Added conditional mandatory properties
 * Properties can have a type
 * Check not yet implemented methods
-* Added all domain exception
 * Minor fixes and improvements
 * Expose property accessor
 * Adoption of semantic versioning
 * Added a service to export ValueObject in JSON
-* Added conditional mandatory properties
+* Removed some code duplications
+* Removed phpdocumentor support
 
 # Release 0.1.5